added confirmation prompt on deletion and finalization of repos

+++ frontend/src/app/components/edit-form/edit-form.component.ts
@@ -1,6 +1,6 @@
-import { Component, OnInit } from "@angular/core";
-import { Repository } from "src/app/models/Repository";
-import { BackendAPIService } from "../../services/backend-api.service";
+import {Component, OnInit} from "@angular/core";
+import {Repository} from "src/app/models/Repository";
+import {BackendAPIService} from "../../services/backend-api.service";
 
 @Component({
   selector: "app-delete-form",
@@ -38,38 +38,42 @@ export class EditFormComponent implements OnInit {
   }
 
   onDeleteRepo(e) {
-    this.loadingDelete = true;
-    const repoId = e.target.value;
-    this.backendApiService.deleteRepository(repoId).subscribe(
-      repository => {
-        this.repositories = this.repositories.filter(repo => repo.id != repoId);
-        this.loadingDelete = false;
-        alert("Repository successfully deleted");
-      },
-      err => {
-        this.errorMessage = err.message;
-        setTimeout(() => {
-          this.errorMessage = "";
-        }, 5000);
-      }
-    );
+    if (window.confirm("Do you really want to delete the repository?")) {
+      this.loadingDelete = true;
+      const repoId = e.target.value;
+      this.backendApiService.deleteRepository(repoId).subscribe(
+        repository => {
+          this.repositories = this.repositories.filter(repo => repo.id != repoId);
+          this.loadingDelete = false;
+          alert("Repository successfully deleted");
+        },
+        err => {
+          this.errorMessage = err.message;
+          setTimeout(() => {
+            this.errorMessage = "";
+          }, 5000);
+        }
+      );
+    }
   }
 
   onFinalizeRepo(e) {
-    this.loadingFinalize = true;
-    const repoId = e.target.value;
-    this.backendApiService.finalizeRepository(repoId).subscribe(
-      repository => {
-        this.repositories = this.repositories.filter(repo => repo.id != repoId);
-        this.loadingFinalize = false;
-        alert("Repository successfully Finalized");
-      },
-      err => {
-        this.errorMessage = err.message;
-        setTimeout(() => {
-          this.errorMessage = "";
-        }, 5000);
-      }
-    );
+    if (window.confirm("Do you really want to finalize the configuration?")) {
+      this.loadingFinalize = true;
+      const repoId = e.target.value;
+      this.backendApiService.finalizeRepository(repoId).subscribe(
+        repository => {
+          this.repositories = this.repositories.filter(repo => repo.id != repoId);
+          this.loadingFinalize = false;
+          alert("Repository successfully Finalized");
+        },
+        err => {
+          this.errorMessage = err.message;
+          setTimeout(() => {
+            this.errorMessage = "";
+          }, 5000);
+        }
+      );
+    }
   }
 }
